Refactor TreeNode to its own submodule
