+++ test/services/network/tracing/tree/test_tree_node_tracker.py
@@ -4,7 +4,7 @@
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at https://mozilla.org/MPL/2.0/.
 from zepben.evolve import Breaker
-from zepben.evolve.services.network.tracing.tree.downstream_tree import TreeNode
+from zepben.evolve.services.network.tracing.tree.tree_node import TreeNode
 from zepben.evolve.services.network.tracing.tree.tree_node_tracker import TreeNodeTracker
